// src/integration/gas-station/business/dtos/gas-station.dto.ts
export interface GasStationDTO {
    id: string;
    name: string;
    brand: string;
    street: string;
    houseNumber: number;
    place: string;
    postCode: number;
    isOpen: boolean;
    distance: number;
    lat: number;
    lng: number;
    dieselPrice: number;
    e5Price: number;
    e10Price: number;
    price: number;
}